shared process - move service to electron-sandbox
