+++ src/vs/platform/ipc/electron-sandbox/sharedProcessService.ts
