+++ src/vs/workbench/workbench.desktop.main.ts
@@ -88,10 +88,8 @@ import 'vs/workbench/services/diagnostics/electron-browser/diagnosticsService';
 import { registerSingleton } from 'vs/platform/instantiation/common/extensions';
 import { ITunnelService } from 'vs/platform/remote/common/tunnel';
 import { TunnelService } from 'vs/workbench/services/remote/electron-browser/tunnelServiceImpl';
-import { ISharedProcessService, SharedProcessService } from 'vs/platform/ipc/electron-browser/sharedProcessService';
 
 registerSingleton(ITunnelService, TunnelService);
-registerSingleton(ISharedProcessService, SharedProcessService, true);
 
 //#endregion
